resource_manager/host_manager/requirements.txt
pika
pyyaml